Some work to store runtime information was made. Menus created in one level, system get all information from runtimeinfo

+++ stl_task_2/CommandList.cpp
@@ -45,7 +45,7 @@ void CommandList::Clear()
     commands.clear();
 }
 
-int CommandList::Size()
+size_t CommandList::Size()
 {
     return commands.size();
 }
